chore: move back global types to each project
