+++ packages/client/src/components/CompleteForm.tsx
@@ -5,9 +5,9 @@ import CameraIcon from '../assets/camera.svg'
 import CloseIcon from '../assets/close.svg'
 import VideoIcon from '../assets/video.svg'
 import { H4, Paragraph } from './Typography'
-import { IFormData } from '../../../../types'
 import { StyledBtn } from './Button'
 import { devices } from '../utils/devices'
+import { IFormData } from '../utils/types'
 
 const TextArea = styled.textarea`
   width: 94%;
