+++ packages/client/src/utils/types.ts
@@ -1,6 +1,24 @@
 export interface APIRequest {
   url: string
 }
+export interface IFormData {
+  text?: string
+  image?: any
+  video?: any
+}
+
+export type ApiExceptionType = {
+  message: string
+}
+
+export interface ErrorReportType {
+  id?: string
+  place: string
+  room: string
+  area: string
+  object: string
+  complete: IFormData
+}
 
 export enum RoomEnum {
   'KITCHEN' = 'Kök',
