+++ packages/server/src/services/types.ts
@@ -5,6 +5,24 @@ export interface Room {
   shared: boolean
 }
 
+export type ApiExceptionType = {
+  message: string
+}
+
+export interface IFormData {
+  text?: string
+  image?: any
+  video?: any
+}
+export interface ErrorReportType {
+  id?: string
+  place: string
+  room: string
+  area: string
+  object: string
+  complete: IFormData
+}
+
 export const InventoryClassification: { [index: string]: string } = {
   i1: 'Anslags-/namn-tavla',
   i2: 'Avloppssystem',
