--- types/index.ts
@@ -1,18 +1,0 @@
-export interface IFormData {
-  text?: string
-  image?: any
-  video?: any
-}
-
-export interface ErrorReportType {
-  id?: string
-  place: string
-  room: string
-  area: string
-  object: string
-  complete: IFormData
-}
-
-export type ApiExceptionType = {
-  message: string
-}
